# [Performance][Spec Decode] Optimize ngram lookup performance