After benchmarking the performance of ngram in vllm, it seems that the proposal time is longer than expected. The main reason is that there are (1) CPU <-> GPU communication when building the ngram lookup table. (2) Building the ngram contains many small kernels (duration < 5 microseconds) as show below:
<img width="783" alt="Screenshot 2024-10-13 at 11 07 44 PM" src="https://github.com/user-attachments/assets/5ef7d0e4-31a0-43ff-9a19-5a3ded9be009">

Zoom in the propose time:
<img width="723" alt="Screenshot 2024-10-13 at 11 09 55 PM" src="https://github.com/user-attachments/assets/95ac97c5-65b3-426e-9556-5fd45a433e5a">

The PR tries to
(1) perform lookup operation on CPU.
(2) trigger CPU <-> GPU communication only when there is a match in lookup.

Some performance numbers on a single H100:
input_len: 550, output_len: 150
I changed the prompt to try different system efficiency (which might include the number of CPU <-> GPU sync).
| System efficiency | propose time before this PR | propose time after this PR | end2end latency before this PR | end2end latency after this PR |
| -- | -- | -- | -- | -- |
| 0.31 | 4.4ms | 2.2ms | 6.4s | 5.6s |
| 0.63 | 3.3ms | 1.5ms | 3.8s | 3.2s |
| 0.80 | 2.6 ms | 1.5ms | 3.0s | 2.6s |


input_len: 2048, output_len: 150
| System efficiency | propose time before this PR | propose time after this PR | end2end latency before this PR | end2end latency after this PR |
| -- | -- | -- | -- | -- |
| 0.30 | 6.00ms | 4.54ms | 9.83s | 9.25s |
| 0.63 | 2.90ms | 2.70ms | 5.84s | 5.45s |