@LiuXiaoxuanPKU 2024-10-14
> btw does this approach still have speedup if the prompt length is much longer? I'm just thinking about the trade off between CPU-GPU sync overhead and (maybe) slower CPU computation.

Yeah will do more benchmarks here